Move LogContext to src/logger.ts
